Python BlackJack Project from Python Bootcamp
This BlackJack project is based on 1 player only against the Dealer.
The logic is very simple and does not support split or double down.  Dealer has to hit at 15.  It covers conditions like Bust, Push (Tie) etc...  Good exercise for using Dice, List manipulation and pop() function/Stack as well as simple double for loop.  This is using inheritance and OOP structure with instantiation of object.